requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
decorator==4.4.2
docopt==0.6.2
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
joblib==1.1.0
moviepy==1.0.3
mp3-tagger==1.0
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
pipreqs==0.4.11
praw==7.6.0
prawcore==2.3.0
proglog==0.1.10
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.64.0
update-checker==0.18.0
urllib3==1.26.11
websocket-client==1.3.3
yarg==0.1.9